### 2024-02-02 #6316 @editorialbot
[PRE REVIEW]: pyRTC: An open-source Python solution for kHz real-time control of adaptive optics systems
**Submitting author:** @jacotay7 (<a href="http://orcid.org/0000-0002-6356-567X">Jacob Taylor</a>)
**Repository:** https://github.com/jacotay7/pyRTC
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
