### 2024-02-07 44aa653
Update utils.py to enable macOS compatability
