### 2024-02-07 24f3935
Delete examples/scao/pywfs_example_OOPAO_pipeline.py
OOPAOInterface doesn't yet support hardware launcher
