### 2024-03-05 0b9d23c
Added example of very basic OOPAO set-up
