+++ examples/scao/simple_OOPAO_example/simpleParamFile.py
@@ -1,0 +1,85 @@
+import numpy as np
+from os import path
+from OOPAO.tools.tools  import createFolder
+
+def initializeParameterFile():
+    # initialize the dictionaries
+    param = dict()
+    
+    ###%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%% ATMOSPHERE PROPERTIES %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+    
+    param['r0'                   ] = 0.1                                            # value of r0 in the visible in [m]
+    param['L0'                   ] = 30                                             # value of L0 in the visible in [m]
+    param['fractionnalR0'        ] = [1]                        # Cn2 profile
+    param['windSpeed'            ] = [10]                               # wind speed of the different layers in [m.s-1]
+    param['windDirection'        ] = [0]                             # wind direction of the different layers in [degrees]
+    param['altitude'             ] = [0]                    # altitude of the different layers in [m]
+                    
+    ###%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%% M1 PROPERTIES %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+    
+    param['diameter'             ] = 1                                             # diameter in [m]
+    param['nSubaperture'         ] = 4                                             # number of PWFS subaperture along the telescope diameter
+    param['nPixelPerSubap'       ] = 8                                              # sampling of the PWFS subapertures
+    param['resolution'           ] = param['nSubaperture']*param['nPixelPerSubap']  # resolution of the telescope driven by the PWFS
+    param['sizeSubaperture'      ] = param['diameter']/param['nSubaperture']        # size of a sub-aperture projected in the M1 space
+    param['samplingTime'         ] = 1/1000                                         # loop sampling time in [s]
+    param['centralObstruction'   ] = 0.                                          # central obstruction in percentage of the diameter
+    param['nMissingSegments'     ] = 0                                              # number of missing segments on the M1 pupil
+    param['m1_reflectivity'      ] = 1                                              # reflectivity of the 798 segments
+          
+    ###%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%% NGS PROPERTIES %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+    
+    param['magnitude'            ] = 4                                              # magnitude of the guide star
+    param['opticalBand'          ] = 'R'                                            # optical band of the guide star
+    param['sourceBand'          ] = 'K'
+
+    ###%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%% DM PROPERTIES %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+    param['nActuator'            ] = param['nSubaperture']+1                        # number of actuators 
+    param['mechanicalCoupling'   ] = 0.45
+    param['isM4'                 ] = False       
+    param['dm_coordinates'       ] = None                                        # tag for the deformable mirror class
+
+    # mis-registrations                                                             
+    param['shiftX'               ] = 0                                              # shift X of the DM in pixel size units ( tel.D/tel.resolution ) 
+    param['shiftY'               ] = 0                                              # shift Y of the DM in pixel size units ( tel.D/tel.resolution )
+    param['rotationAngle'        ] = 0                                              # rotation angle of the DM in [degrees]
+    param['anamorphosisAngle'    ] = 0                                              # anamorphosis angle of the DM in [degrees]
[…]
