### 2024-03-09 #6316 @arfon
@andrewpaulreeves @matthewtownson @jfsauvage @joao-aveiro :wave: would any of you be willing to review this submission for JOSS? The submission under consideration is _pyRTC: An open-source Python solution for kHz real-time control of adaptive optics systems_

Many thanks
Arfon
