### 2024-03-11 #6316 @joseph-long
@dfm and I were just talking about this submission at lunch! Have never reviewed for JOSS before but would be happy to give it a go, perhaps with @joao-aveiro's help
