### 2024-03-11 #6316 @arfon
> @dfm and I were just talking about this submission at lunch! Have never reviewed for JOSS before but would be happy to give it a go, perhaps with @joao-aveiro's help

Whoah! Hi @joseph-long! Thanks for agreeing to help out here.

> Hi @arfon! Thank you for the invitation. I can review this submission on the basis of general OSS development quality and compliance with the journal's guidelines, as well as basic AO and RTC scientific pertinence and correctness. Nevertheless, due to my limited expertise on the latter, I think it would be beneficial to include another reviewer more knowledgeable on these scientific fields.

@joao-aveiro – thanks so much!

> I too would be happy to review. I'm working on a related RTC project in Julia so the subject matter is currently front of mind.

🙇 thank you @sefffal.

I'll go ahead and add you all now and start the review ⚡
