+++ SHARP_LAB/softRTC.py
@@ -1,18 +1,27 @@
 # %% Imports
 import numpy as np
 import matplotlib.pyplot as plt
-# from pyRTC.hardware.ximeaWFS import *
-from pyRTC.WavefrontSensor import *
+from pyRTC.hardware.ximeaWFS import *
+# from pyRTC.WavefrontSensor import *
 from pyRTC.hardware.ALPAODM import *
 from pyRTC.utils import *
 from pyRTC.hardware.SpinnakerScienceCam import *
+from pyRTC.SlopesProcess import *
+from pyRTC.Loop import *
+#%% CLEAR SHMs
+# shms = ["wfs", "wfsRaw", "signal", "signal2D", "wfc", "wfc2D", "psfShort", "psfLong"]
+# clear_shms(shms)
 # %% Load Config
 conf = read_yaml_file("/home/whetstone/pyRTC/SHARP_LAB/config.yaml")
 # %% Launch WFS
 confWFS = conf["wfs"]
-wfs = WavefrontSensor(conf=confWFS)
-time.sleep(4)
+wfs = XIMEA_WFS(conf=confWFS)
+time.sleep(0.5)
 wfs.start()
+# %% Launch slopes
+slopes = SlopesProcess(conf=conf)
+slopes.start()
+time.sleep(0.5)
 # %% Launch WFC
 confWFC = conf["wfc"]
 wfc = ALPAODM(conf=confWFC)
@@ -21,33 +30,115 @@ wfc.start()
 # %% Launch PSF
 confPSF = conf["psf"]
 psf = spinCam(conf=confPSF)
+time.sleep(0.5)
 psf.start()
-time.sleep(2)
-#  %% Visualize IM
-IM = np.load("/home/whetstone/pyRTC/SHARP_LAB/IM.npy")
-plt.imshow(IM, aspect="auto")
-plt.show()
 
-# for row in np.linspace(0,10,5).astype(int):
-#     x = signal2D(IM[:,row], wfs.layout)
-#     plt.imshow(x)
-#     plt.show()
+# %% Launch loop
+loop = Loop(conf=conf)
+time.sleep(1)
 
-# %% Flatten
-FLAT = np.load(confWFC['flatFile'])
-wfc.setFlat(FLAT)
+# %% Recalibrate
+
+# Darks
+
+input("Sources Off?")
+wfs.takeDark()
+wfs.darkFile = "/home/whetstone/pyRTC/SHARP_LAB/dark.npy"
+wfs.saveDark()
+time.sleep(1)
+psf.takeDark()
+psf.darkFile = "/home/whetstone/pyRTC/SHARP_LAB/psfDark.npy"
+psf.saveDark()
+input("Sources On?")
+input("Is Atmosphere Out?")
+
+slopes.computeImageNoise()
+slopes.refSlopesFile =  ""
+slopes.loadRefSlopes()
+slopes.takeRefSlopes()
+slopes.refSlopesFile = "/home/whetstone/pyRTC/SHARP_LAB/ref.npy"
+slopes.saveRefSlopes()
+
 wfc.flatten()
+psf.takeModelPSF()
+psf.modelFile = "/home/whetstone/pyRTC/SHARP_LAB/modelPSF.npy"
+psf.saveModelPSF()
+
+#  STANDARD IM
+loop.IMMethod = "push-pull"
+loop.pokeAmp = 0.03
+loop.numItersIM = 100
+loop.IMFile = "/home/whetstone/pyRTC/SHARP_LAB/IM.npy"
+wfc.flatten()
+loop.computeIM()
+loop.saveIM()
+wfc.flatten()
+time.sleep(1)
+
+input("Is Atmosphere In?")
[…]
