+++ pyRTC/RTC.py
@@ -2,8 +2,15 @@
 #Import pyRTC classes
 from pyRTC.Pipeline import *
 from pyRTC.utils import *
-
+import os
+os.chdir("/home/whetstone/pyRTC/pyRTC")
 RECALIBRATE = False
+
+# %% Clear SHMs
+# from pyRTC.Pipeline import clear_shms
+# shm_names = ["signal"]
+# shm_names = ["wfs", "wfsRaw", "wfc", "wfc2D", "signal", "signal2D", "psfShort", "psfLong"] #list of SHMs to reset
+# clear_shms(shm_names)
 
 # %% IMPORTS
 config = '/home/whetstone/pyRTC/SHARP_LAB/config.yaml'
@@ -21,7 +28,7 @@ slopes = hardwareLauncher("./SlopesProcess.py", config, N+2)
 slopes.launch()
 
 # %% Launch PSF Cam
-psfCam = hardwareLauncher("./hardware/SpinnakerScienceCam.py", config, N+3)
+psfCam = hardwareLauncher("./hardware/SpinnakerScienceCam.py", config, N+10)
 psfCam.launch()
 
 # %% Launch Loop Class
@@ -38,16 +45,28 @@ if RECALIBRATE == True:
     ##### slopes.setProperty("offsetY", 3)
 
     input("Sources Off?")
-    psfCam.run("takeDark")
-    psfCam.run("saveDark")
     wfs.run("takeDark")
+    wfs.setProperty("darkFile", "/home/whetstone/pyRTC/SHARP_LAB/dark.npy")
     wfs.run("saveDark")
+    time.sleep(1)
+    psfCam.run("takeDark")
+    psfCam.setProperty("darkFile", "/home/whetstone/pyRTC/SHARP_LAB/psfDark.npy")
+    psfCam.run("saveDark")
     input("Sources On?")
+    input("Is Atmosphere Out?")
 
+    slopes.run("computeImageNoise")
     slopes.run("takeRefSlopes")
     slopes.setProperty("refSlopesFile", "/home/whetstone/pyRTC/SHARP_LAB/ref.npy")
     slopes.run("saveRefSlopes")
 
+    wfc.run("flatten")
+    psfCam.run("takeModelPSF")
+    psfCam.setProperty("modelFile", "/home/whetstone/pyRTC/SHARP_LAB/modelPSF.npy")
+    psfCam.run("saveModelPSF")
+
+    #  STANDARD IM
+    loop.setProperty("IMMethod", "push-pull")
     loop.setProperty("pokeAmp", 0.03)
     loop.setProperty("numItersIM", 100)
     loop.setProperty("IMFile", "/home/whetstone/pyRTC/SHARP_LAB/IM.npy")
@@ -57,23 +76,141 @@ if RECALIBRATE == True:
     wfc.run("flatten")
     time.sleep(1)
 
-# %%
-loop.setProperty("numDroppedModes", 20)
-loop.run("computeCM")
+    input("Is Atmosphere In?")
+    #  DOCRIME OL
+    loop.setProperty("IMMethod", "docrime")
+    loop.setProperty("delay", 1)
+    loop.setProperty("pokeAmp", 2e-2)
+    loop.setProperty("numItersIM", 30000)
+    loop.setProperty("IMFile", "/home/whetstone/pyRTC/SHARP_LAB/docrime_IM.npy")
+    wfc.run("flatten")
+    loop.run("computeIM")
+    loop.run("saveIM")
+    wfc.run("flatten")
[…]

[…]
