### 2024-04-17 d9d86ce
update to the simple sim for custom gym env
