+++ pyRTC/hardware/LoopGymInterface.py
@@ -15,27 +15,48 @@ class LoopGymInterface(RealTimeGymInterface):
     def __init__(self, loop):
         self.loop = loop
         self.initialized = False
+        self.control = self.get_default_action()
+        self.flat2D = np.zeros((self.loop.wfc2D_width, self.loop.wfc2D_height))
+
+
+    def merge_slope_cmd(self, slope, cmd):
+
+        if slope.shape[1] >= cmd.shape[1]:
+
+            pad_size = slope.shape[1] - cmd.shape[1]
+
+            padded_cmd = np.pad(cmd, ((0, 0), (0, pad_size)), mode='constant', constant_values=0)
+
+            return np.concatenate((slope, padded_cmd), axis=0)
+        
+        else:
+            pad_size = cmd.shape[1] - slope.shape[1]
+
+            padded_slopes = np.pad(slope, ((0, 0), (0, pad_size)), mode='constant', constant_values=0)
+
+            return np.concatenate((padded_slopes, cmd), axis=0)
 
 
 
     def get_action_space(self):
         numActuators = self.loop.confWFC['numModes']
 
-        high_cmd = np.ones(numActuators) * np.inf
-        low_cmd = np.ones(numActuators) * -np.inf
+        return spaces.Box(low=-1, high=1, shape=(numActuators, ) , dtype=np.float32)
 
-        return spaces.Box(low=low_cmd, high=high_cmd, dtype=np.float32)
+
 
     def get_observation_space(self):
 
-        high_wfs = np.ones((self.loop.slope_width, self.loop.slope_height)) * np.inf
-        low_wfs = np.ones((self.loop.slope_width, self.loop.slope_height)) * -np.inf
+        if self.loop.slope_height >= self.loop.wfc2D_height:
 
-        slope_obs = spaces.Box(low=low_wfs, high=high_wfs, dtype=np.float32)
+            pad_size = 0
+    
+        else:
+            pad_size = self.loop.wfc2D_height - self.loop.slope_height
 
-        cmd_obs = self.get_action_space()
 
-        return spaces.Tuple([cmd_obs, slope_obs])
+        return spaces.Box(low=-1, high=1, shape=(self.loop.slope_width + self.loop.wfc2D_width, self.loop.slope_height + pad_size), dtype=np.float32)
+
 
 
     def get_default_action(self):
@@ -45,6 +66,9 @@ class LoopGymInterface(RealTimeGymInterface):
     def send_control(self, control):
 
         control[self.loop.numActiveModes:] = 0
+
+        self.control = control
+        
         self.loop.wfcShm.write(control)
 
 
@@ -54,28 +78,38 @@ class LoopGymInterface(RealTimeGymInterface):
             self.initialized = True
 
         # Flatten the mirror
-        self.loop.wfcShm.write(self.get_default_action())
+        flat_cmd = self.get_default_action()
 
-        # Read actuator positions
-        act_pos = self.loop.wfcShm.read()
+        self.send_control(flat_cmd)
+
+        # Get flat dm map
+        dm = self.flat2D
             
         # Read current slope map
-        slopes = self.loop.slopesShm.read()
+        slopes = self.loop.slopesShm.read_noblock()
 
-        return [np.array(act_pos, dtype='float32'),
-                np.array(slopes, dtype='float32')], {}
+        obs = self.merge_slope_cmd(slopes, dm)
+
+    
+        return np.array(obs), {}
 
 
     def get_obs_rew_terminated_info(self):
         
         # Read actuator positions
-        act_pos = self.loop.wfcShm.read()
+        act_pos = self.control
+
[…]
