+++ examples/scao/simple_OOPAO_example/simple_OOPAO_config.yaml
@@ -33,7 +33,7 @@ slopes:
 wfc:
   name: "OOPAOWFC"
   numActuators: 21
-  numModes: 21
+  numModes: 18
   flatFile: ""
   saveFile: "flat.npy"
   m2cFile:  ""
