### 2024-04-22 #6466 @sefffal
Hi @jacotay7, I have a few initial questions & comments.
First off, as you may know, there is an existing python library called `pyRTC` which has some slight overlap in functionality. Would you consider at all picking a slightly different name for the project, to minimize confusion?

Second, I notice that the documentation is a little sparse so far. There is a great quick-start notebook that would tell me how to set up a basic system.

But a typical AO system would be expected to need a fair amount of application-specific customization. As it stands, I'm not sure the example notebook would provide enough information for me to know how to create a multi-stage pipeline, add new wavefront sensors, or support for a DM, for instance.

Would you consider writing an additional page explaining the architecture of the software? If this documentation is already written and I have overlooked it, please disregard!
