### 2024-04-24 b95a7d9
Added support for remote science camera process
