+++ REVOLT/softRTC.py
@@ -7,14 +7,12 @@ from pyRTC.hardware.ALPAODM import *
 from pyRTC.hardware.fliCBlueOneWFS import *
 from pyRTC.SlopesProcess import *
 from pyRTC.Loop import *
-
 # %% Load Config
 conf = read_yaml_file("/home/revoltuser/pyRTC/REVOLT/config.yaml")
-
+RECALIBRATE = False
 # %% Clear SHM
 # shm_names = ["wfs", "wfsRaw", "wfc", "wfc2D", "signal", "signal2D"] #list of SHMs to reset
 # clear_shms(shm_names)
-
 # %% Launch WFC
 confWFC = conf["wfc"]
 BASIS = "MODAL"
@@ -28,121 +26,273 @@ elif BASIS == "MODAL":
 wfc = ALPAODM(conf=confWFC)
 time.sleep(0.5)
 wfc.start()
-
-
 # %% Launch WFS
 confWFS = conf["wfs"]
 wfs = FliCBlueOneWFS(conf=confWFS)
 wfs.start()
-
 # %% Launch Slopes
 slopes = SlopesProcess(conf=conf)
 slopes.start()
-# %% Monitor WFS
-slopes.shwfsContrast = 7 # set to 0 if not dark subtracting, slowly dial up to ideal
-print(np.mean(wfs.read()))
-slopes.offsetX = 0
-slopes.offsetY = 0
+# %% Launch loop
+loop = Loop(conf=conf)
+time.sleep(1)
 
-for i in range(1):
-    im = wfs.read()
-    plt.imshow(im > np.mean(im) * slopes.shwfsContrast)
-    plt.colorbar()
-    plt.show()
-    plt.imshow(slopes.read())
-    plt.colorbar()
-    plt.show()
-    time.sleep(0.5)
+# %% Recalibrate
 
+# Darks
+if RECALIBRATE:
+    input("Sources Off?")
+    # input("FAKE DARK PLEASE ACKNOWLEDGE")
+    # wfs.dark = np.zeros_like(wfs.dark)
+    wfs.takeDark()
+    wfs.darkFile = "/home/revoltuser/pyRTC/REVOLT/dark.npy"
+    wfs.saveDark()
+    time.sleep(1)
+    input("Sources On?")
+    input("Is Atmosphere Out?")
 
+    slopes.computeImageNoise()
+    slopes.refSlopesFile =  ""
+    slopes.loadRefSlopes()
+    slopes.takeRefSlopes()
+    slopes.refSlopesFile = "/home/revoltuser/pyRTC/REVOLT/ref.npy"
+    slopes.saveRefSlopes()
+    wfc.flatten()
 
-# %% Calibration
-#Take Dark
-# input("Is Lamp OFF???")
-# wfs.takeDark()
-# wfs.darkFile = "/home/revoltuser/pyRTC/REVOLT/dark.npy"
-# wfs.saveDark()
-
-#Take Reference Slopes
-# input("Is Lamp ON???")
-# wfc.flatten()
-# slopes.refSlopesFile = "/home/revoltuser/pyRTC/REVOLT/refSlopes.npy"
-# slopes.takeRefSlopes()
-# slopes.saveRefSlopes()
-
-# %% Loop
-loop = Loop(conf)
-
-
-# %% Compute IM
-if BASIS == "MODAL":
-    loop.pokeAmp = 0.03
-else:
+    #  STANDARD IM
+    loop.IMMethod = "push-pull"
     loop.pokeAmp = 0.02
-loop.numItersIM = 100
-loop.IMFile =  "./IM.npy"
+    loop.numItersIM = 100
[…]
