### 2024-04-29 #6466 @jacotay7
> Hi @jacotay7, I have a few initial questions & comments. First off, as you may know, there is an existing python library called `pyRTC` which has some slight overlap in functionality. Would you consider at all picking a slightly different name for the project, to minimize confusion?

Yeah, I know that there is already a pyRTC on pip, I think I will change the name once I decide things are far enough along to be put on pip, and I will switch to the name pyAORTC or similar. For now, I think all users will be cloning the repo individually and I think the likelihood of conflict with the exiting repository is very small.

> Second, I notice that the documentation is a little sparse so far. There is a great quick-start notebook that would tell me how to set up a basic system.
>
> But a typical AO system would be expected to need a fair amount of application-specific customization. As it stands, I'm not sure the example notebook would provide enough information for me to know how to create a multi-stage pipeline, add new wavefront sensors, or support for a DM, for instance.
>
> Would you consider writing an additional page explaining the architecture of the software? If this documentation is already written and I have overlooked it, please disregard!

Sorry about this, I am in the process of getting a guide written with details on the implementation. I have been keeping a full config and RTC example for my lab system in the main branch to compensate for the lack of this guide. I will update you when this guide is avaliable.

On a more general note, this package is very hard to test since it is a hardware interfacing code, which is why I put together a simulation example in the first place. The simulation examples cannot really fully capture the utility of the program the same way a hardware implementation can, without some imagination on the part of the user. If anyone has any suggestions for how to bridge that gap for the review let me know.
