+++ pyRTC/Pipeline.py
@@ -107,6 +107,13 @@ class ImageSHM:
         arr = np.copy(self.arr)
         return arr
     
+    def read_timeout(self, timeout):
+        start = time.time()
+        while not self.checkNew() and (time.time() - start) < timeout:
+            time.sleep(1e-5)
+        arr = np.copy(self.arr)
+        return arr
+    
     def read_noblock(self):
         arr = np.copy(self.arr)
         return arr
@@ -117,14 +124,19 @@ class ImageSHM:
     def checkNew(self):
         
         if self.areData:
-            metadata = np.copy(self.metadata)
-            if metadata[1] != self.lastReadTime:
-                self.lastReadTime = metadata[1]
+            # metadata = np.copy(self.metadata)
+            if self.metadata[1] != self.lastReadTime:
+                self.markSeen()
+                # self.lastReadTime = self.metadata[1]
                 return True
         else: #If we are just reading a meta data object directly
             return True
         return False
     
+    def markSeen(self):
+        self.lastReadTime = self.metadata[1]
+        return
+
     def updateMetadata(self):
         
         metadata = np.zeros_like(self.metadata)
@@ -149,13 +161,15 @@ def clear_shms(names):
 
 class hardwareLauncher:
 
-    def __init__(self, hardwareFile, configFile, port) -> None:
+    def __init__(self, hardwareFile, configFile, port, timeout=None) -> None:
         self.hardwareFile = hardwareFile
         self.command = ["python", hardwareFile, "-c", f"{configFile}", "-p", f"{port}"]
         self.running = False
         # Client configuration
         self.host = '127.0.0.1'  # localhost
         self.port = port
+        self.timeout = timeout
+
         return
     
     def launch(self):
@@ -168,7 +182,7 @@ class hardwareLauncher:
             self.processSocket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
             print(f"Waiting for Process at {self.host}:{self.port}")
             connected = False
-            restTime = 1
+            restTime = 2
             while not connected:
                 time.sleep(restTime)
                 try:
@@ -178,6 +192,9 @@ class hardwareLauncher:
                 except Exception as e:
                     print(f"Connection failed: {e}")
                     print("Retrying in {} seconds...".format(restTime))
+
+            if isinstance(self.timeout,float) or isinstance(self.timeout,int):
+                self.processSocket.settimeout(self.timeout)
 
             print("Connected")
 
@@ -195,7 +212,7 @@ class hardwareLauncher:
         message = {"type": "set", "property": property, "value": value}
         return self.writeAndRead(message)
 
-    def run(self, function, *args):
+    def run(self, function, *args, timeout = None):
         message = {"type": "run", "function": function}
         for i, arg in enumerate(args):
             message[f"arg_{i+1}"] = arg
@@ -228,8 +245,12 @@ class hardwareLauncher:
         return
     
     def read(self):
-        reply = self.processSocket.recv(4096).decode()
-        return json.loads(reply)
+        try:
+            reply = self.processSocket.recv(4096).decode()
+            return json.loads(reply)
+        except socket.timeout:
+            return -1
+        
     
 
 class Listener:
@@ -241,12 +262,16 @@ class Listener:
         self.host = '127.0.0.1'  # localhost
         self.port = port
 
-        # Create a socket object
-        server_socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+        server_socket = bind_socket(self.host, self.port)
 
-        print(f"{hardware.name}: Binding to {self.host}:{self.port}")
[…]
