+++ SHARP_LAB/softRTC.py
@@ -40,53 +40,53 @@ time.sleep(1)
 # %% Recalibrate
 
 # Darks
+if False:
+    input("Sources Off?")
+    wfs.takeDark()
+    wfs.darkFile = "/home/whetstone/pyRTC/SHARP_LAB/dark.npy"
+    wfs.saveDark()
+    time.sleep(1)
+    psf.takeDark()
+    psf.darkFile = "/home/whetstone/pyRTC/SHARP_LAB/psfDark.npy"
+    psf.saveDark()
+    input("Sources On?")
+    input("Is Atmosphere Out?")
 
-input("Sources Off?")
-wfs.takeDark()
-wfs.darkFile = "/home/whetstone/pyRTC/SHARP_LAB/dark.npy"
-wfs.saveDark()
-time.sleep(1)
-psf.takeDark()
-psf.darkFile = "/home/whetstone/pyRTC/SHARP_LAB/psfDark.npy"
-psf.saveDark()
-input("Sources On?")
-input("Is Atmosphere Out?")
+    slopes.computeImageNoise()
+    slopes.refSlopesFile =  ""
+    slopes.loadRefSlopes()
+    slopes.takeRefSlopes()
+    slopes.refSlopesFile = "/home/whetstone/pyRTC/SHARP_LAB/ref.npy"
+    slopes.saveRefSlopes()
 
-slopes.computeImageNoise()
-slopes.refSlopesFile =  ""
-slopes.loadRefSlopes()
-slopes.takeRefSlopes()
-slopes.refSlopesFile = "/home/whetstone/pyRTC/SHARP_LAB/ref.npy"
-slopes.saveRefSlopes()
+    wfc.flatten()
+    psf.takeModelPSF()
+    psf.modelFile = "/home/whetstone/pyRTC/SHARP_LAB/modelPSF.npy"
+    psf.saveModelPSF()
 
-wfc.flatten()
-psf.takeModelPSF()
-psf.modelFile = "/home/whetstone/pyRTC/SHARP_LAB/modelPSF.npy"
-psf.saveModelPSF()
+    #  STANDARD IM
+    loop.IMMethod = "push-pull"
+    loop.pokeAmp = 0.03
+    loop.numItersIM = 100
+    loop.IMFile = "/home/whetstone/pyRTC/SHARP_LAB/IM.npy"
+    wfc.flatten()
+    loop.computeIM()
+    loop.saveIM()
+    wfc.flatten()
+    time.sleep(1)
 
-#  STANDARD IM
-loop.IMMethod = "push-pull"
-loop.pokeAmp = 0.03
-loop.numItersIM = 100
-loop.IMFile = "/home/whetstone/pyRTC/SHARP_LAB/IM.npy"
-wfc.flatten()
-loop.computeIM()
-loop.saveIM()
-wfc.flatten()
-time.sleep(1)
-
-input("Is Atmosphere In?")
-#  DOCRIME OL
-loop.IMMethod = "docrime"
-loop.delay = 3
-loop.pokeAmp = 2e-2
-loop.numItersIM = 10000
-loop.IMFile = "/home/whetstone/pyRTC/SHARP_LAB/docrime_IM.npy"
-wfc.flatten()
-loop.computeIM()
-loop.saveIM()
-wfc.flatten()
-time.sleep(1)
+    input("Is Atmosphere In?")
+    #  DOCRIME OL
+    loop.IMMethod = "docrime"
+    loop.delay = 3
+    loop.pokeAmp = 2e-2
+    loop.numItersIM = 10000
+    loop.IMFile = "/home/whetstone/pyRTC/SHARP_LAB/docrime_IM.npy"
+    wfc.flatten()
[…]
