### 2024-05-08 45ef02e
Added NCPA optimizer, general support for optimizers. Removed old RTCs
