+++ SHARP_LAB/softRTC.py
@@ -12,7 +12,7 @@ from pyRTC.Loop import *
 # shms = ["wfs", "wfsRaw", "signal", "signal2D", "wfc", "wfc2D", "psfShort", "psfLong"]
 # clear_shms(shms)
 # %% Load Config
-conf = read_yaml_file("/home/whetstone/pyRTC/SHARP_LAB/config.yaml")
+conf = read_yaml_file("/home/whetstone/pyRTC/SHARP_LAB/config_SR.yaml")
 # %% Launch WFS
 confWFS = conf["wfs"]
 wfs = XIMEA_WFS(conf=confWFS)
@@ -32,7 +32,6 @@ confPSF = conf["psf"]
 psf = spinCam(conf=confPSF)
 time.sleep(0.5)
 psf.start()
-
 # %% Launch loop
 loop = Loop(conf=conf)
 time.sleep(1)
@@ -43,11 +42,11 @@ time.sleep(1)
 if False:
     input("Sources Off?")
     wfs.takeDark()
-    wfs.darkFile = "/home/whetstone/pyRTC/SHARP_LAB/dark.npy"
+    wfs.darkFile = "/home/whetstone/pyRTC/SHARP_LAB/calib/dark.npy"
     wfs.saveDark()
     time.sleep(1)
     psf.takeDark()
-    psf.darkFile = "/home/whetstone/pyRTC/SHARP_LAB/psfDark.npy"
+    psf.darkFile = "/home/whetstone/pyRTC/SHARP_LAB/calib/psfDark_SR.npy"
     psf.saveDark()
     input("Sources On?")
     input("Is Atmosphere Out?")
@@ -56,42 +55,42 @@ if False:
     slopes.refSlopesFile =  ""
     slopes.loadRefSlopes()
     slopes.takeRefSlopes()
-    slopes.refSlopesFile = "/home/whetstone/pyRTC/SHARP_LAB/ref.npy"
+    slopes.refSlopesFile = "/home/whetstone/pyRTC/SHARP_LAB/calib/ref.npy"
     slopes.saveRefSlopes()
 
     wfc.flatten()
     psf.takeModelPSF()
-    psf.modelFile = "/home/whetstone/pyRTC/SHARP_LAB/modelPSF.npy"
+    psf.modelFile = "/home/whetstone/pyRTC/SHARP_LAB/calib/modelPSF.npy"
     psf.saveModelPSF()
 
     #  STANDARD IM
     loop.IMMethod = "push-pull"
     loop.pokeAmp = 0.03
     loop.numItersIM = 100
-    loop.IMFile = "/home/whetstone/pyRTC/SHARP_LAB/IM.npy"
+    loop.IMFile = "/home/whetstone/pyRTC/SHARP_LAB/calib/IM.npy"
     wfc.flatten()
     loop.computeIM()
     loop.saveIM()
     wfc.flatten()
     time.sleep(1)
 
-    input("Is Atmosphere In?")
-    #  DOCRIME OL
-    loop.IMMethod = "docrime"
-    loop.delay = 3
-    loop.pokeAmp = 2e-2
-    loop.numItersIM = 10000
-    loop.IMFile = "/home/whetstone/pyRTC/SHARP_LAB/docrime_IM.npy"
-    wfc.flatten()
-    loop.computeIM()
-    loop.saveIM()
-    wfc.flatten()
-    time.sleep(1)
+    # input("Is Atmosphere In?")
+    # #  DOCRIME OL
+    # loop.IMMethod = "docrime"
+    # loop.delay = 3
+    # loop.pokeAmp = 2e-2
+    # loop.numItersIM = 10000
[…]
