+++ SHARP_LAB/RTC.py
@@ -3,38 +3,42 @@
 from pyRTC.Pipeline import *
 from pyRTC.utils import *
 import os
-os.chdir("/home/whetstone/pyRTC/pyRTC")
+os.chdir("/home/whetstone/pyRTC/SHARP_LAB")
 RECALIBRATE = False
 
 # %% Clear SHMs
-from pyRTC.Pipeline import clear_shms
-shm_names = ["signal"]
-shm_names = ["wfs", "wfsRaw", "wfc", "wfc2D", "signal", "signal2D", "psfShort", "psfLong"] #list of SHMs to reset
-clear_shms(shm_names)
+# from pyRTC.Pipeline import clear_shms
+# shm_names = ["signal"]
+# shm_names = ["wfs", "wfsRaw", "wfc", "wfc2D", "signal", "signal2D", "psfShort", "psfLong"] #list of SHMs to reset
+# clear_shms(shm_names)
 
 # %% IMPORTS
-config = '/home/whetstone/pyRTC/SHARP_LAB/config.yaml'
+config = '/home/whetstone/pyRTC/SHARP_LAB/config_SR.yaml'
 N = np.random.randint(3000,6000)
 # %% Launch DM
-wfc = hardwareLauncher("./hardware/ALPAODM.py", config, N)
+wfc = hardwareLauncher("../pyRTC/hardware/ALPAODM.py", config, N)
 wfc.launch()
 
 # %% Launch WFS
-wfs = hardwareLauncher("./hardware/ximeaWFS.py", config, N+1)
+wfs = hardwareLauncher("../pyRTC/hardware/ximeaWFS.py", config, N+1)
 wfs.launch()
 
 # %% Launch slopes
-slopes = hardwareLauncher("./SlopesProcess.py", config, N+2)
+slopes = hardwareLauncher("../pyRTC/SlopesProcess.py", config, N+2)
 slopes.launch()
 
 # %% Launch PSF Cam
-psfCam = hardwareLauncher("./hardware/SpinnakerScienceCam.py", config, N+10)
+psfCam = hardwareLauncher("../pyRTC/hardware/SpinnakerScienceCam.py", config, N+10)
 psfCam.launch()
 
 # %% Launch Loop Class
-loop = hardwareLauncher("./Loop.py", config, N+4)
-# loop = hardwareLauncher("./hardware/predictLoop.py", config)
+loop = hardwareLauncher("../pyRTC/Loop.py", config, N+4)
+# loop = hardwareLauncher("../pyRTC/hardware/predictLoop.py", config)
 loop.launch()
+
+# %% NCAP OPTIMIZER
+optim  = hardwareLauncher("../pyRTC/hardware/NCPAOptimizer.py", config, N+4)
+optim.launch()
 
 # %% Calibrate
 
@@ -46,52 +50,63 @@ if RECALIBRATE == True:
 
     input("Sources Off?")
     wfs.run("takeDark")
-    wfs.setProperty("darkFile", "/home/whetstone/pyRTC/SHARP_LAB/dark.npy")
+    wfs.setProperty("darkFile", "/home/whetstone/pyRTC/SHARP_LAB/calib/dark.npy")
     wfs.run("saveDark")
     time.sleep(1)
     psfCam.run("takeDark")
-    psfCam.setProperty("darkFile", "/home/whetstone/pyRTC/SHARP_LAB/fullPsfDark.npy")
+    psfCam.setProperty("darkFile", "/home/whetstone/pyRTC/SHARP_LAB/calib/psfDark.npy")
     psfCam.run("saveDark")
     input("Sources On?")
     input("Is Atmosphere Out?")
 
[…]
