+++ docs/source/components/wfc.rst
@@ -1,0 +1,27 @@
+.. wfs:
+
+.. automodule:: pyRTC.WavefrontCorrector
+
+
+Wavefront Corrector
+====================
+
+Add description
+
+
+Example
+-------
+
+Describe Example
+
+.. code-block:: python
+
+  from pyRTC.WavefrontCorrector import *
+
+
+Parameters
+----------
+
+.. autoclass:: WavefrontCorrector
+  :members:
+  :inherited-members:
