+++ docs/source/guides/getting_started.rst
@@ -1,0 +1,6 @@
+.. getting started
+
+Getting Started
+===============
+
+TODO
