+++ docs/source/components/modulator.rst
@@ -1,0 +1,27 @@
+.. modulator:
+
+.. automodule:: pyRTC.Modulator
+
+
+Modulator
+================
+
+Add description
+
+
+Example
+-------
+
+Describe Example
+
+.. code-block:: python
+
+  from pyRTC.Modulator import *
+
+
+Parameters
+----------
+
+.. autoclass:: Modulator
+  :members:
+  :inherited-members:
