+++ SHARP_LAB/config.yaml
@@ -13,10 +13,17 @@ loop:
   # AIdevice: "cuda:0"
   # AIburnIn: 50000
   # AIresidualLength: 250000
+  pGain: 0.316
+  iGain: 0.0612
+  dGain: 0.0481
+  controlLimits: [-0.1, 0.1]
+  integralLimits: [-0.1, 0.1]
+  derivativeFilter: 0.1
   functions:
   # - standardIntegrator
-  - standardIntegratorPOL
+  # - standardIntegratorPOL
   # - leakyIntegrator
+  - pidIntegrator
 
 
 wfs:
@@ -100,14 +107,14 @@ psf:
   exposure: 300
   gain: 0
   bitDepth: 16
-  left: 920 #768
-  top: 436 #248
-  width: 128 #640
-  height: 128 #640
+  left: 920
+  top: 436
+  width: 128
+  height: 128
   darkCount: 2000
   darkFile: "/home/whetstone/pyRTC/SHARP_LAB/calib/psfDark.npy"
   modelFile: "/home/whetstone/pyRTC/SHARP_LAB/calib/modelPSF.npy"
-  integration: 5
+  integration: 1000
   affinity: 6
   functions:
   - expose
@@ -115,14 +122,22 @@ psf:
   - computeStrehl
 
 optimizer:
-  name: "NCPA"
-  logdir: "ncpa_optimizer_tensorboard"
-  numSteps: 200
-  startMode: 0
-  endMode: 94
-  correctionMag: 0.003
-  numReads: 10
-  affinity: 10
+  ncpa:
+    name: "NCPA"
+    numSteps: 100
+    startMode: 0
+    endMode: 20
+    correctionMag: 0.003
+    numReads: 1
+    affinity: 10
+  pid:
+    name: "PID"
+    numSteps: 100
+    maxPGain: 0.5
+    maxIGain: 0.1
+    maxDGain: 0.1
+    numReads: 1
+    affinity: 10
 
 # modulator:
 #   name: 'E-727'
