### 2024-05-21 dfcd5f5
``	modified:   pyRTC/SlopesProcess.py
