### 2024-05-29 2ad12df
Added some more documentation for wfs
