+++ pyRTC/WavefrontSensor.py
@@ -10,23 +10,116 @@ from numba import jit
 from sys import platform
 
 class WavefrontSensor(pyRTCComponent):
+    """
+    A pyRTCComponent which represents a Wavefront Sensor (camera). This is a general class which is 
+    reponsible for all components of wavefront sensing which are common to all wavefront sensors. This
+    class should be used by defining a child class held in pyRTC.hardware, which overwrites
+    the relevant functions which actual hardware connectivity code. The child class can call its parent
+    implementations in order to make use of the code which sets the relevant parameters, write to shared
+    memory, etc... or they can overwrite them completely.
 
-    def __init__(self, conf) -> None:
+    Config
+    ------
+    name : str
+        The name of the wavefront sensor. Default "wavefrontSensor"
+    width : int
+        The width of the wavefront sensor image. Required.
+    height : int
+        The width of the wavefront sensor image.  Required.      
+    darkCount : int
+        Number of dark frames to average. Default 1000.
+    darkFile : str
+        Path to the dark frame file. Default, empty string.
 
-        self.name = conf["name"]
-        self.imageShape = (conf["width"], conf["height"])
+    Attributes
+    ----------
+    imageShape : tuple
+        The shape of the image (width, height).
+    imageRawDType : data-type
+        The data type for raw image.
+    imageDType : data-type
+        The data type for processed image.
+    imageRaw : ImageSHM
+        Shared memory object for raw image.
+    image : ImageSHM
+        Shared memory object for processed image.
+    data : ndarray
+        Array to store raw image data.
+    dark : ndarray
+        Array to store dark frame data.
+    affinity : int
+        The affinity configuration.
+    roiWidth : int
+        Width of the region of interest.
+    roiHeight : int
+        Height of the region of interest.
+    roiLeft : int
+        Left coordinate of the region of interest.
+    roiTop : int
+        Top coordinate of the region of interest.
+    exposure : float
+        Exposure time.
+    binning : int
+        Binning factor.
+    gain : float
+        Gain setting.
+    bitDepth : int
+        Bit depth of the image.
+
+    Methods
+    -------
+    setRoi(roi)
+        Sets the region of interest.
+    setExposure(exposure)
+        Sets the exposure time.
+    setBinning(binning)
+        Sets the binning factor.
+    setGain(gain)
+        Sets the gain.
+    setBitDepth(bitDepth)
+        Sets the bit depth.
+    expose()
+        Writes the current image data to shared memory.
+    read()
+        Reads the processed image data from shared memory.
+    takeDark()
+        Captures and sets the dark frame.
+    setDark(dark)
+        Sets the dark frame.
+    saveDark(filename='')
+        Saves the dark frame to a file.
+    loadDark(filename='')
+        Loads the dark frame from a file.
+    plot()
+        Plots the current image data.
+    """
+
+    def __init__(self, conf: dict) -> None:
+        """
+        Constructs all the necessary attributes for the wavefront sensor object.
+
+        Parameters
+        ----------
+        conf : dict
+            Configuration dictionary for the wavefront sensor. Typically it will just be
+            the "wfs" section of a pyRTC config.
+        """
+
+        self.name = setFromConfig(conf, "name", "wavefrontSensor")
+        self.width = conf["width"]
+        self.height = conf["height"]
+        self.darkCount = setFromConfig(conf, "darkCount", 1000)
+        self.darkFile = setFromConfig(conf, "darkFile", "")
+
+        self.imageShape = (self.width, self.height)
         self.imageRawDType = np.uint16
         self.imageDType = np.int32
-        
+
[…]
