### 2024-06-04 aa65822
Post-Run Commit. Added Telemetry service, optical gain GUI, changes to Optimizers, optical gain integrators.
