+++ pyRTC/Pipeline.py
@@ -161,23 +161,24 @@ def clear_shms(names):
 
 class hardwareLauncher:
 
-    def __init__(self, hardwareFile, configFile, port, timeout=None) -> None:
+    def __init__(self, hardwareFile, configFile, port, remoteProcess=False, timeout=None) -> None:
         self.hardwareFile = hardwareFile
         self.command = ["python", hardwareFile, "-c", f"{configFile}", "-p", f"{port}"]
         self.running = False
         # Client configuration
         self.host = '127.0.0.1'  # localhost
         self.port = port
+        self.remoteProcess = remoteProcess
         self.timeout = timeout
 
         return
     
     def launch(self):
         if not self.running:
-            print(f"Launching Process: {self.hardwareFile}")
-            self.process = Popen(self.command,stdin=PIPE,stdout=PIPE, text=True, bufsize=1)
-            self.running = True
-
+            if not self.remoteProcess:
+                print(f"Launching Process: {self.hardwareFile}")
+                self.process = Popen(self.command,stdin=PIPE,stdout=PIPE, text=True, bufsize=1)
+                
             # Create a socket object
             self.processSocket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
             print(f"Waiting for Process at {self.host}:{self.port}")
@@ -193,6 +194,7 @@ class hardwareLauncher:
                     print(f"Connection failed: {e}")
                     print("Retrying in {} seconds...".format(restTime))
 
+            self.running = True
             if isinstance(self.timeout,float) or isinstance(self.timeout,int):
                 self.processSocket.settimeout(self.timeout)
 
@@ -255,11 +257,11 @@ class hardwareLauncher:
 
 class Listener:
 
-    def __init__(self, hardware, port) -> None:
+    def __init__(self, hardware, port, host = '127.0.0.1') -> None:
         self.hardware = hardware
         self.running = True
         self.keyCharacter = '$'
-        self.host = '127.0.0.1'  # localhost
+        self.host = host  # default localhost
         self.port = port
 
         server_socket = bind_socket(self.host, self.port)
