+++ SHARP_LAB/softRTC.py
@@ -12,7 +12,7 @@ from pyRTC.Loop import *
 # shms = ["wfs", "wfsRaw", "signal", "signal2D", "wfc", "wfc2D", "psfShort", "psfLong"]
 # clear_shms(shms)
 # %% Load Config
-conf = read_yaml_file("/home/whetstone/pyRTC/SHARP_LAB/config_SR.yaml")
+conf = read_yaml_file("/home/whetstone/pyRTC/SHARP_LAB/config.yaml")
 # %% Launch WFS
 confWFS = conf["wfs"]
 wfs = XIMEA_WFS(conf=confWFS)
@@ -94,7 +94,7 @@ loop.IMFile = "/home/whetstone/pyRTC/SHARP_LAB/calib/IM.npy"
 loop.loadIM()
 loop.numDroppedModes = 40
 loop.computeCM()
-loop.setGain(0.01)
+loop.setGain(0.1)
 loop.leakyGain = 1e-2
 
 # %% Start Loop
@@ -134,43 +134,42 @@ loop.numDroppedModes = 20
 loop.computeCM()
 # %%
 SIM = np.load("/home/whetstone/pyRTC/SHARP_LAB/calib/sprint_IM_nomisreg_valid.npy").reshape(94, -1).T
-bench_converter_SL = (np.linalg.pinv(SIM) @ loop.IM)[:,:21]
-bench_converter_LS = (loop.CM @ SIM)[:,:21]
-bench_converter_NP = np.eye(94)[:,:21]
+bench_converter_SL = (np.linalg.pinv(SIM) @ loop.IM) #[:,:21]
+bench_converter_LS = (loop.CM @ SIM) #[:,:21]
+bench_converter_NP = np.eye(94) #[:,:21]
 # %% Tip-Tilt-Focus-Sweep
 from tqdm import tqdm
 
-folder = "/home/whetstone/Downloads/robin-may-10"
+folder = "/home/whetstone/Downloads/robin-may-23"
 
-
-# numModes = 10
-# startMode = 0
-# endMode = wfc.numModes - 1
-# filelist = ['cnnx2_phase.npy', 'cnnx4_phase.npy', 'linx2_phase.npy', 'linx4_phase.npy']
-
-filelist = ['cnnx2_phase', 'cnnx4_phase','linx2_phase', 'linx4_phase', 'cnnx8n3_phase','linx8n3_phase']
+# filelist = ['cnnx2_phase', 'cnnx4_phase','linx2_phase', 'linx4_phase', 'cnnx8n3_phase','linx8n3_phase']
+# filelist = ['_zern_cnn_x2_n4_phase', '_zern_cnn_x4_n4_phase', '_zern_cnn_x8_n4_phase', '_fixzern_lin_x2_n4_phase', '_fixzern_lin_x4_n4_phase', '_fixzern_lin_x8_n4_phase']
 # N = 3
 
-numModes = 11
-modelist = [0, 0.25, 0.5, 1., 1.5] #np.linspace(-RANGE, RANGE, numModes) #.astype(int)
+# numModes = 11
+# powerlist = [0, 0.25, 0.5, 1., 2., 4., 8., 16] #np.linspace(-RANGE, RANGE, numModes) #.astype(int)
+powerlist = np.linspace(-10, 10, 101)
+numModes = len(powerlist)
 slopecorrect = 0.0021
 
-for k, bench_converter in enumerate([bench_converter_SL, bench_converter_LS, bench_converter_NP]):
-    bc = ["SL", "LS", "NP"][k]
-    for ff in filelist:
+for k, bench_converter in enumerate([bench_converter_NP]): #bench_converter_SL, bench_converter_LS, bench_converter_NP]):
[…]
