+++ pyRTC/utils.py
@@ -2,22 +2,112 @@ import yaml
 import sys
 import select
 import os 
+from astropy.io import fits
 from subprocess import PIPE, Popen
 import numpy as np
 import psutil
 from scipy.ndimage import median_filter, gaussian_filter
 import socket
+from datetime import datetime
 
 NP_DATA_TYPES = [
     np.int8, np.int16, np.int32, np.int64,
     np.uint8, np.uint16, np.uint32, np.uint64,
-    np.float16, np.float32, np.float64, #np.float128,  # np.float128 availability depends on the system
-    np.complex64, np.complex128, #np.complex256,       # np.complex256 availability depends on the system
+    np.float16, np.float32, np.float64, np.float128,  # np.float128 availability depends on the system
+    np.complex64, np.complex128, np.complex256,       # np.complex256 availability depends on the system
     np.bool_,
     np.object_,
     np.string_, np.unicode_,
     np.datetime64, np.timedelta64
 ]
+
+def powerLawOG(numModes, k):
+    return (1- (np.arange(numModes)/numModes)**k)
+
+
+def append_to_file(filename, data, dtype=np.float32):
+    """
+    Append a numpy array to a binary file on disk.
+
+    Parameters:
+    filename : str
+        The name of the file to which data will be appended.
+    data : numpy array
+        The numpy array to append to the file.
+    dtype : data-type, optional
+        The desired data-type for the array. Default is np.float32.
+    """
+    if os.path.exists(filename):
+        # If the file exists, append to it
+        with open(filename, 'ab') as f:
+            data.tofile(f)
+    else:
+        # If the file does not exist, create it and write the initial data
+        with open(filename, 'wb') as f:
+            data.tofile(f)
+
+def generate_circular_aperture_mask(N, R, ratio):
+    """
+    Generates a binary mask of size NxN with a circular aperture of radius R and a central obscuration of radius r.
+    
+    Parameters:
+    N (int): The size of the mask (NxN).
+    R (float): The radius of the outer circular aperture.
+    ratio (float): The ratio of the inner obscuration radius to the outer radius (r/R).
+
+    Returns:
+    numpy.ndarray: Binary mask with the circular aperture.
+    """
+    r = R * ratio
+    x = np.linspace(-N/2, N/2, N)
+    xx, yy = np.meshgrid(x,x)
+    mask = (xx**2 + yy**2 <= R**2) 
+    if r > 0:
+        mask &= (xx**2 + yy**2 >= r**2)
+    return mask.astype(bool)
+
+def load_data(filename, dtype=None):
+    if filename.endswith('.npy'):
+        data = np.load(filename)
+    elif filename.endswith('.fits'):
+        with fits.open(filename) as hdul:
+            data = hdul[0].data
+    else:
+        raise ValueError("Unsupported file format. Please provide a .npy or .fits file.")
+    
+    if dtype is not None:
+        return data.astype(dtype)
+    return data
+
+def generate_filepath(base_dir='.', prefix='file', extension='.dat'):
+    """
+    Generate a file path based on the current date and time.
+
+    Parameters:
+    base_dir : str
+        The base directory where the file will be saved.
+    prefix : str
+        The prefix for the file name.
+    extension : str
+        The file extension.
+
+    Returns:
+    str
+        The generated file path.
+    """
+    # Get the current date and time
+    current_time = datetime.now()
+
+    # Format the date and time
+    timestamp = current_time.strftime('%Y%m%d_%H%M%S')
+
+    # Construct the file name
+    filename = f"{prefix}_{timestamp}{extension}"
+
+    # Construct the full file path
+    filepath = os.path.join(base_dir, filename)
+
+    return filepath
 
 def get_tmp_filepath(file_path):
     """
