### 2024-06-06 3c82c9b
Fixed the issues with signal2D in slopes process for pyramids
