### 2024-06-07 4b46925
Fixed hardware init to check from SDKs, added some defaults and small fixed to WFC
