### 2024-06-07 1273863
Initial control system for SUPER PAOWER. Just added communication, initialization
