### 2024-06-07 c3a0f7d
Debugging why serial port isn't working on Linux
