### 2024-06-11 2e9ce6b
Added more documentation for loop components.
