+++ pyRTC/ScienceCamera.py
@@ -8,7 +8,86 @@ import numpy as np
 import matplotlib.pyplot as plt
 
 class ScienceCamera(pyRTCComponent):
+    """
+    A pyRTCComponent which represents a Science Camera. This is a general class which is 
+    reponsible for all components of imaging which are common to all Science Cameras. This
+    class should be used by defining a child class held in pyRTC.hardware, which overwrites
+    the relevant functions which actual hardware connectivity code. The child class can call its parent
+    implementations in order to make use of the code which sets the relevant parameters, write to shared
+    memory, etc... or they can overwrite them completely. See hardware/SpinnakerScienceCam.py for an example.
 
+    Config
+    ------
+    name : str
+        Name of the camera.
+    width : int
+        Width of the image. Required.
+    height : int
+        Height of the image. Required.
+    darkCount : int
+        Number of dark frames to average. Required.
+    integration : int
+        Integration length. Required.
+    darkFile : str, optional
+        File to save the dark frames. Default is "".
+    modelFile : str, optional
+        File to save the model PSF. Default is "".
+
+    Attributes
+    ----------
+    name : str
+        Name of the camera.
+    imageShape : tuple
+        Shape of the image.
+    imageRawDType : type
+        Data type of the raw image.
+    imageDType : type
+        Data type of the image.
+    psfLongDtype : type
+        Data type of the long exposure PSF.
+    psfShort : ImageSHM
+        Shared memory object for the short exposure PSF.
+    psfLong : ImageSHM
+        Shared memory object for the long exposure PSF.
+    strehlShm : ImageSHM
+        Shared memory object for the Strehl ratio.
+    tipTiltShm : ImageSHM
+        Shared memory object for the tip-tilt.
+    data : numpy.ndarray
+        Data array for the image.
+    dark : numpy.ndarray
+        Dark frame.
+    darkCount : int
+        Number of dark frames to average.
+    darkFile : str
+        File to save the dark frames.
+    model : numpy.ndarray
+        Model PSF.
+    modelFile : str
+        File to save the model PSF.
+    strehl_ratio : float
+        Strehl ratio.
+    peak_dist : float
+        Peak distance.
+    integrationLength : int
+        Integration length.
+    roiWidth : int
+        Width of the region of interest.
+    roiHeight : int
+        Height of the region of interest.
+    roiLeft : int
+        Left coordinate of the region of interest.
+    roiTop : int
+        Top coordinate of the region of interest.
+    exposure : int
+        Exposure time.
+    binning : int
+        Binning factor.
+    gain : int
+        Gain setting.
+    bitDepth : int
+        Bit depth setting.
+    """
     def __init__(self, conf) -> None:
 
         self.name = conf["name"]
@@ -38,6 +117,14 @@ class ScienceCamera(pyRTCComponent):
         super().__init__(conf)
     
     def setRoi(self, roi):
+        """
+        Set the region of interest (ROI).
+
+        Parameters
+        ----------
+        roi : tuple
+            Tuple containing (width, height, left, top) of the ROI.
+        """
         self.roiWidth = roi[0]
         self.roiHeight = roi[1]
         self.roiLeft = roi[2]
@@ -45,43 +132,111 @@ class ScienceCamera(pyRTCComponent):
         return
 
     def setExposure(self, exposure):
+        """
+        Set the exposure time.
+
+        Parameters
+        ----------
+        exposure : int
+            Exposure time to set.
+        """
         self.exposure = exposure
         return
     
     def setBinning(self, binning):
+        """
+        Set the binning factor.
+
+        Parameters
+        ----------
+        binning : int
+            Binning factor to set.
+        """
         self.binning = binning
         return
     
     def setGain(self, gain):
[…]
