+++ pyRTC/WavefrontCorrector.py
@@ -24,7 +24,82 @@ def ModaltoZonalWithFlat(correction=np.array([],dtype=np.float32),
     return M2C@correction + flat
 
 class WavefrontCorrector(pyRTCComponent):
+    """
+    A pyRTCComponent which represents a Wavefront Corrector (DM, SLM, other). This is a general class which is 
+    reponsible for all components of wavefront correcting which are common to all wavefront correctors. This
+    class should be used by defining a child class held in pyRTC.hardware, which overwrites
+    the relevant functions which actual hardware connectivity code. The child class can call its parent
+    implementations in order to make use of the code which sets the relevant parameters, write to shared
+    memory, etc... or they can overwrite them completely. See hardware/ALPAODM.py for an example.
 
+    Config
+    ------
+    name : str
+        Name of the wavefront corrector.
+    numActuators : int
+        Number of actuators. Required.
+    numModes : int
+        Number of modes. Required.
+    affinity : str
+        Affinity setting.
+    m2cFile : str
+        Path to the mode-to-command file.
+    floatingInfluenceRadius : int, optional
+        Radius for floating influence. Default is 1.
+    frameDelay : int, optional
+        Frame delay. Default is 0.
+    saveFile : str, optional
+        File to save the shape. Default is "wfcShape.npy".
+
+    Attributes
+    ----------
+    name : str
+        Name of the wavefront corrector.
+    numActuators : int
+        Number of actuators.
+    numModes : int
+        Number of modes.
+    affinity : str
+        Affinity setting.
+    m2cFile : str
+        Path to the mode-to-command file.
+    correctionVector : ImageSHM
+        Correction vector.
+    correctionVector2D : ImageSHM or None
+        2D correction vector for display.
+    flat : numpy.ndarray
+        Initial flat shape.
+    flatModal : numpy.ndarray
+        Flat shape in modal basis.
+    currentShape : numpy.ndarray
+        Current shape.
+    actuatorStatus : numpy.ndarray
+        Status of each actuator.
+    index_map : numpy.ndarray or None
+        Index map for actuators.
+    floatingInfluenceRadius : int
+        Radius for floating influence.
+    floatMatrix : numpy.ndarray
+        Floating actuator matrix.
+    frameDelay : int
+        Frame delay.
+    saveFile : str
+        File to save the shape.
+    layout : numpy.ndarray or None
+        Layout of the actuators.
+    M2C : numpy.ndarray
+        Mode-to-command matrix.
+    f_M2C : numpy.ndarray
+        Floating mode-to-command matrix.
+    C2M : numpy.ndarray
+        Command-to-mode matrix.
+    currentCorrection : numpy.ndarray
+        Current correction vector.
+    shapeBuffer : numpy.ndarray
+        Buffer for shapes with frame delay.
+    correctionVector2D_template : numpy.ndarray
+        Template for the 2D correction vector.
+    """
     def __init__(self, conf) -> None:
 
         self.name = conf["name"]
@@ -61,10 +136,26 @@ class WavefrontCorrector(pyRTCComponent):
         return
 
     def setFlat(self, flat):
+        """
+        Set the flat shape.
+
+        Parameters
+        ----------
+        flat : numpy.ndarray
+            Flat shape to set.
+        """
         self.flat = flat
         return
 
     def setLayout(self, layout):
+        """
+        Set the layout of the actuators.
+
+        Parameters
+        ----------
+        layout : numpy.ndarray or None
+            Layout of the actuators. Is converted to boolean if not already.
+        """
         self.layout = layout
         if isinstance(self.layout, np.ndarray):
             self.layout = self.layout > 0
@@ -78,7 +169,14 @@ class WavefrontCorrector(pyRTCComponent):
[…]
