### 2024-06-12 1f02fef
Added instructions for building documentation, added some code examples for wfs, wfc, etc ... (not completed yet)
