+++ pyRTC/Optimizer.py
@@ -12,9 +12,53 @@ import time
 import optuna
 
 class Optimizer(pyRTCComponent):
+    """
+    The Optimizer component for is for general optimization tasks 
+    in pyRTC. This class should be used by defining a child class held in pyRTC.hardware, 
+    which overwrites the relevant functions which actual hardware connectivity code. 
+    The child class can call its parent implementations in order to make use of the code 
+    which sets the relevant parameters, write to shared memory, etc... or they can overwrite 
+    them completely. See hardware/NCPAOptimizer.py for an example.
+
+    This class is designed to perform optimization using Optuna's 
+    CmaEsSampler and manages the optimization process including 
+    the application of optimum values and trial management.
+
+    :param conf: Configuration dictionary containing necessary parameters.
+
+    **Config Parameters**:
+    - **numSteps** (*int*): The number of steps/trials to perform during optimization. Default is 100.
+
+    Attributes
+    ----------
+    name : str
+        Name of the optimizer component.
+    study : optuna.Study
+        The study object from Optuna, initialized with a CmaEsSampler.
+    numSteps : int
+        Number of steps/trials to perform during optimization.
+
+    Methods
+    -------
+    objective():
+        Defines the objective function for the optimization.
+    optimize():
+        Performs the optimization process.
+    applyOptimum():
+        Applies the optimum values obtained from the optimization process.
+    applyTrial(trial):
+        Applies a given trial.
+    applyNext():
+        Requests and applies the next trial from the study.
+    """
+
 
     def __init__(self, conf) -> None:
-        
+        """
+        Initializes the Optimizer with the given configuration.
+
+        :param conf: Configuration dictionary containing necessary parameters.
+        """
         self.name = "Optimizer"
         self.study = optuna.create_study(direction='maximize', 
                                          sampler=optuna.samplers.CmaEsSampler())
@@ -25,68 +69,55 @@ class Optimizer(pyRTCComponent):
         return
     
     def objective(self):
+        """
+        Defines the objective function for the optimization.
+
+        This method should be overridden by subclasses to provide the 
+        specific objective function for the optimization task.
+
+        :return: The objective value to be optimized.
+        """
         return
 
     def optimize(self):
+        """
+        Performs the optimization process.
+
+        This method runs the optimization process using the defined objective 
+        function and the number of steps specified in the configuration.
+        """
         self.study.optimize(self.objective, 
                             n_trials=self.numSteps)
         self.applyOptimum()
         return
     
     def applyOptimum(self):
+        """
+        Applies the optimum values obtained from the optimization process.
+
+        This method should be implemented to apply the optimal parameters 
+        found during the optimization to the system or component.
+        """
         return
     
     def applyTrial(self, trial):
+        """
+        Applies a given trial.
+
+        :param trial: The trial object containing the parameters to be applied.
+        """
         return
     
     def applyNext(self):
+        """
+        Requests and applies the next trial from the study.
+
+        This method obtains the next trial from the study and applies it 
+        using the applyTrial method.
+        """
         self.applyTrial(self.study.ask())
         return
 
-    def gradientDescent(self, numSteps, eps, rate):
-
-        #Vector to hold performance
-        loss = np.empty(numSteps)
-        #Vector to hold current NCPA
-        ncpaVec = np.zeros(self.env.num_correction_modes, 
-                           dtype=self.loop.IM.dtype)
-        
-        #Arrays to hold derivatives
-        pds = np.zeros((numSteps,self.env.num_correction_modes))
-        pd = np.zeros(self.env.num_correction_modes)
-
[…]
