### 2024-06-22 #6466 @joseph-long
Hi all, checking in and acknowledging my tardiness on this! It occurs to me that @jacotay7 and I could have discussed this at SPIE last week :)

I agree with @sefffal's comments so far. It would be great to get any planned package name change out the door, as well as some more detailed documentation (which @jacotay7 alluded to in his talk!)
