+++ pyRTC/Loop.py
@@ -93,7 +93,17 @@ class Loop(pyRTCComponent):
         self.IMMethod = setFromConfig(self.confLoop, "IMMethod", "push-pull") 
         self.IMFile = setFromConfig(self.confLoop, "IMFile", "")
         
-
+        self.clDocrime = False     
+        self.numItersDC = 0   
+        tmp2 = self.flat.copy()
+        tmp2 = tmp2.reshape(tmp2.size,1) 
+        tmp = self.nullSignal.copy()
+        tmp = tmp.reshape(tmp.size,1)
+        self.docrimeCross = np.zeros_like(tmp@tmp2.T)
+        self.docrimeAuto = np.zeros_like(tmp2@tmp2.T)
+        self.docrimeBuffer = np.zeros((1+self.delay, *tmp2.shape), 
+                                dtype=self.wfcDType)
+        
         """
         Terms for PID integrator
         """
@@ -133,7 +143,7 @@ class Loop(pyRTCComponent):
             #Plus amplitude
             correction[i] = self.pokeAmp
             #Post a new shape to be made
-            self.wfcShm.write(correction)
+            self.sendToWfc(correction)
             #Add some delay to ensure one-to-one
             time.sleep(self.hardwareDelay)
             #Burn the first new image since we were moving the DM during the exposure
@@ -147,7 +157,7 @@ class Loop(pyRTCComponent):
             #Minus amplitude
             correction[i] = -self.pokeAmp
             #Post a new shape to be made
-            self.wfcShm.write(correction)
+            self.sendToWfc(correction)
             #Add some delay to ensure one-to-one
             time.sleep(self.hardwareDelay)
             #Burn the first new image since we were moving the DM during the exposure
@@ -174,39 +184,35 @@ class Loop(pyRTCComponent):
         correction = correction.reshape(correction.size,1)
 
         #Have a history of corrections
-        corrections = np.zeros((1+self.delay, *correction.shape), dtype=correction.dtype)
+        # corrections = np.zeros((1+self.delay, *correction.shape), dtype=correction.dtype)
 
         #Get an initial slope reading to set shapes
         slopes = self.nullSignal.copy()
         slopes = slopes.reshape(slopes.size,1)
-        cross = np.zeros_like(slopes@correction.T)
-        auto = np.zeros_like(correction@correction.T)
+        self.docrimeCross = np.zeros_like(self.docrimeCross)
+        self.docrimeAuto = np.zeros_like(self.docrimeAuto)
+
 
         for i in range(self.numItersIM):
             #Compute new random shape
             correction = np.random.uniform(-self.pokeAmp,self.pokeAmp,correction.size).astype(correction.dtype).reshape(correction.shape)
-            #If we are in Closed Loop
-            if self.running:
-                #Read the current shape of the WFC and add our perturbation ontop
-                correction = self.wfcShm.read_noblock() + correction
-
-            #Send our new pertubation to the WFC
-            self.wfcShm.write(correction.reshape(corrShapeWFC))
-            #Move old shapes back in history
-            corrections[:-1] = corrections[1:]
-            #Add new correction
-            corrections[-1] = correction
-
+            
             #Get current WFS response
+            #I put this first to match CL case
             slopes = self.signalShm.read().reshape(slopes.shape)
-        
-            #Correlate Current response with old correction by delay time
-            cross += slopes@corrections[0].T
-            auto += corrections[0]@corrections[0].T
 
-        cross /= self.numItersIM 
-        auto /= self.numItersIM
-        self.IM = cross@np.linalg.inv(auto) 
+            #Send random shape to mirror
[…]
