### 2024-07-08 #6466 @sefffal
@arfon I think we're not blocked on the review, just waiting for some fixes on @jacotay7's side--in particular I'd like some more documentation so that I can test out the software for myself (one of the review criteria).

Also, I want to note that there are currently no automated tests present in the repository (either that run in CI, or can be run manually). I'm not sure this is essential for this kind of project but it is part of the standard checklist above.

So perhaps we can pause things here and @jacotay7 can comment when he's ready for us to take another look.
