### 2024-07-22 281d529
Added power meter support. Initial testing on device
