+++ SUPERPAOWER/softRTC.py
@@ -1,28 +1,125 @@
 # %% Imports
+import sys
+orig_stdout = sys.stdout
 import numpy as np
 import matplotlib.pyplot as plt
 from pyRTC import *
 from pyRTC.hardware import *
 from pyRTC.utils import *
 from pyRTC.Pipeline import *
+sys.stdout = orig_stdout
+
 #%% CLEAR SHMs
-shms = ["wfc", "wfc2D"]
-clear_shms(shms)
+# shms = ["wfc", "wfc2D","psfShort", "psfLong"]
+# clear_shms(shms)
 # %% Load Config
 conf = read_yaml_file("./config.yaml")
 # %%
+power = powerMeter(conf["power"])
+power.start()
+# %%
 wfc = SUPERPAOWER(conf["wfc"])
 wfc.start()
+time.sleep(1)
+
+#%% RAMP SINGLE ACTUATOR
+
+ACT = 10
+RAMP_START = 0 # Volt
+RAMP_STOP = 4 # Volt
+RESOLUTION = 30e-3 #Volt
+
+ramp = np.arange(RAMP_START, RAMP_STOP, RESOLUTION)
+response = np.empty_like(ramp)
+
+power.readLong()
+for i, volt in enumerate(ramp):
+    wfc.push(ACT, volt)
+    response[i] = power.readLong()
+
+plt.plot(ramp, response)
+plt.show()
+
+#%% Plot Power Meter Noise
+power.setExposure(100)
+
+response = np.empty(100)
+for i in range(100):
+    response[i] = power.read()[0][0]
+plt.plot(response)
+# plt.ylim(0, 1e-7)
+plt.show()
+print(f"NOISE: {np.std(response)}")
+
 # %%
-wfc.push(0,2)
+VOLTLIM = 10
+resolution = 30e-3 #Volts
+
+numAct = wfc.numActuators
+ramp = np.arange(0, VOLTLIM, resolution)
+response = np.zeros_like(ramp)
+
+responses = np.zeros((numAct, *response.shape))
+
+wfcShm, _, _ = initExistingShm("wfc")
+powerShm, _, _ = initExistingShm("psfShort")
+
+correction = np.zeros_like(wfcShm.read_noblock())
+
+
+# for act in range(4):
+#     correction *= 0
+#     for i, volt in enumerate(ramp):
+#         correction[4*act: 4*(act+1)] = volt
+#         wfcShm.write(correction)
+#         response[i] = powerShm.read()
+#         responses[act] = response
+
+for act in range(7,8):#numAct):
+    correction *= 0
+    for i, volt in enumerate(ramp):
+        correction[act] = volt
+        wfcShm.write(correction)
+        time.sleep(1)
+        response[i] = powerShm.read()
+        responses[act] = response
+        plt.plot(response)
+# np.save("/home/whetstone/SUPERPAOWER/pyRTC/SUPERPAOWER/data/responseCurves", responses)
+
 # %%
-wfc.flatten()
-# %%
-# for i in range(16): 
-#     wfc.push(i,1)
-#     time.sleep(1)
-# %%
-for volt in np.linspace(0,4):
-    wfc.push(0,volt)
-    time.sleep(1)
+responses = np.load("/home/whetstone/SUPERPAOWER/pyRTC/SUPERPAOWER/data/responseCurves.npy")
+plt.figure(figsize=(12,5))
+
+for act in range(responses.shape[0]):
+    if act != 7:
+        continue
+    plt.plot(ramp,responses[act], label = f"Actuator #{act+1}")
+
+plt.legend()
+plt.xlabel("Voltage [V]", size = 18)
+plt.ylabel("Power [W]", size = 18)
[…]
