### 2024-07-26 #6466 @arfon
Got it, thanks @sefffal.

> Also, I want to note that there are currently no automated tests present in the repository (either that run in CI, or can be run manually). I'm not sure this is essential for this kind of project but it is part of the standard checklist above.

Thanks for pointing this out. Automated testing is [not a strict requirement](https://joss.readthedocs.io/en/latest/review_criteria.html#tests) at JOSS but is strongly encouraged.

@jacotay7 – what do you think about adding a test suite?
