### 2024-07-26 #6466 @jacotay7
Hey @arfon, @sefffal

I am still working on adding a readthedocs to the repo with some added documentation and examples. Progress is slow as I am trying to finish up and defend my thesis, but it is still my plan to get around to this.

I think overall, this is a hard piece of code to review because it requires integration into real hardware in order to properly test, so I am really appreciative of the reviewers and feedback so far. I have some data from our lab which coroborates some of the critical assertions in the paper related to things like speed. I appreciate these claims would be difficult to have reviewers verify without assess to hardware.

I think adding tests for some of the more crtical components of the repo is a good idea. Specifically, the slope computations, loop integrators, and focal plane analysis sections of the code would be good to have protections for.  I think for now the documentation and examples is more critical.
