+++ pyRTC/utils.py
@@ -25,7 +25,7 @@ def centroid(array):
     y_indices, x_indices = np.indices(array.shape)
     x_centroid = (x_indices * array).sum() / total
     y_centroid = (y_indices * array).sum() / total
-    return x_centroid, y_centroid
+    return np.array([x_centroid, y_centroid])
 
 def add_to_buffer(buffer, vec):
     buffer[:-1] = buffer[1:]
