+++ pyRTC/RLAO/AOEnv_po4ao.py
@@ -1,0 +1,240 @@
+import gymnasium as gym
+from gymnasium import spaces
+import numpy as np
+from pyRTC.Pipeline import * 
+from pyRTC.utils import *
+import torch
+
+
+class pyRTCEnv(gym.Env):
+
+    metadata = {'render.modes': ['human']}
+
+    def __init__(self, conf):
+        super(pyRTCEnv, self).__init__()
+
+        confLoop = conf["loop"]
+        self.device = setFromConfig(confLoop, 'device', 'cuda:0')
+
+        # DM
+        confDM = conf["wfc"]
+
+        #------------------------------Shared Memory Objects (AO System Components) ---------------------------#
+
+        # Wavefront Corrector
+        self.wfc2DShm, self.wfc2DDims, self.wfc2DDtype = initExistingShm("wfc2D")
+        self.wfcShm, self.wfcDims, self.wfcDtype = initExistingShm("wfc")
+        
+        self.numModes = self.wfcDims[0]
+        self.numActiveModes = self.numModes
+        
+        # Precomputed Slopes
+        self.signalShm, self.signalDims, self.signalDtype = initExistingShm("signal")
+
+        # Strehl Tracker
+        self.strehlShm, self.strehlDims, self.strehlDtype = initExistingShm("strehl")
+
+
+        #------------------------------System Parameters & Interaction----------------------------------#
+
+        # Loop Parameters
+        self.gain  =  setFromConfig(confLoop, "gain", 0.1)
+        self.linearActionNorm = setFromConfig(confLoop, "linearActionNorm", 1)
+        self.actionNorm = setFromConfig(confLoop, "actionNorm", 1000)
+        self.leakyGain =  setFromConfig(confLoop, "leakyGain", 0)
+        self.actionScale = setFromConfig(confLoop, "actionScale", 0.01)
+        self.commandMagnitudeWeight = setFromConfig(confLoop, "commandMagnitudeWeight", 1.0)
+
+        # Calibration Data
+        self.IMFile = setFromConfig(confLoop, "IMFile", "")
+        self.IM = torch.tensor(np.load(self.IMFile), dtype=torch.float32, device=self.device)
+
+        self.M2C_file = setFromConfig(confDM, "m2cFile", "")
+        self.M2C_CL = torch.tensor(np.load(self.M2C_file), dtype=torch.float32, device=self.device)
+        self.C2M = torch.pinverse(self.M2C_CL)
+
+        #Filter (F) projects a shape onto the basis
+        self.F = self.M2C_CL@self.C2M
+
+
+        self.CM = torch.zeros_like(self.IM.T, device=self.device)
+        self.CM[:self.numActiveModes,:] = torch.pinverse(self.IM[:,:self.numActiveModes], 
+                                                         rcond=0)
+        self.CM[self.numActiveModes:,:] = 0
+        self.gCM = self.gain*self.CM
+
+        self.numActuators = setFromConfig(confDM, "numActuators", 97)
+
+        #Actuators across the dm (approximation of squaring the circle)
+        self.nActuator = int(np.round(np.sqrt(self.numActuators/(np.pi/4)),0))
+       
+        if self.numActuators == 97:
+            xx, yy = np.meshgrid(np.arange(11), np.arange(11))
+            layout = np.sqrt((xx - 5)**2 + (yy-5)**2) < 5.5
+        
+        nonzero_indices = torch.nonzero(torch.tensor(layout), as_tuple=True)
+
+        # This will give you two tensors: xvalid and yvalid
+
+        print("TORCH TENSOR THE VALID PIXELS")
+        self.xvalid = nonzero_indices[0].clone().detach()
+        self.yvalid = nonzero_indices[1].clone().detach()
+        
[…]
