+++ pyRTC/RLAO/__init__.py
@@ -1,0 +1,8 @@
+# Import specific functions or classes from submodules
+__all__ = []
+
+try:
+    from .RLLoop import *
+    __all__.append('RLLoop')
+except:
+    print("Check if you have necessary packages install for RLAO support")
