### 2024-08-23 ec2f908
Added gamma controls to science camera
