### 2024-09-26 ca7f2d1
Huge commit from several months of work. Changes to the pipeline to support GPU Shared memory. Some new optimizers, added a easier launch script to launch components from the hardware launcher
