+++ SHARP_LAB/RTC_PyWFS.py
@@ -1,29 +1,21 @@
 # %% IMPORTS
 #Import pyRTC classes
+import matplotlib.pyplot as plt
 from pyRTC.Pipeline import *
 from pyRTC.utils import *
+from pyRTC.hardware import *
 import os
 os.chdir("/home/whetstone/pyRTC/SHARP_LAB")
 RECALIBRATE = False
-
+CLEAR_SHMS =  False
 # %% Clear SHMs
-# from pyRTC.Pipeline import clear_shms
-# shm_names = ["signal"]
-# shm_names = ["wfs", "wfsRaw", "wfc", "wfc2D", "signal", "signal2D", "psfShort", "psfLong"] #list of SHMs to reset
-# clear_shms(shm_names)
-
+if CLEAR_SHMS:
+    shm_names = ["wfs", "wfsRaw", "wfc", "wfc2D", "signal", "signal2D", "psfShort", "psfLong"] #list of SHMs to reset
+    clear_shms(shm_names)
 # %% IMPORTS
 config = '/home/whetstone/pyRTC/SHARP_LAB/config_pywfs.yaml'
 N = np.random.randint(3000,6000)
 folder = "/home/whetstone/pyRTC/SHARP_LAB/calib/"
-
-# %% Launch Modulator
-modulator = hardwareLauncher("../pyRTC/hardware/PIModulator.py", config, N+100)
-modulator.launch()
-
-# %% Launch DM
-wfc = hardwareLauncher("../pyRTC/hardware/ALPAODM.py", config, N)
-wfc.launch()
 
 # %% Launch WFS
 wfs = hardwareLauncher("../pyRTC/hardware/ximeaWFS.py", config, N+1)
@@ -37,6 +29,14 @@ slopes.launch()
 psfCam = hardwareLauncher("../pyRTC/hardware/SpinnakerScienceCam.py", config, N+10)
 psfCam.launch()
 
+# %% Launch DM
+wfc = hardwareLauncher("../pyRTC/hardware/ALPAODM.py", config, N)
+wfc.launch()
+
+# %% Launch Modulator
+modulator = hardwareLauncher("../pyRTC/hardware/PIModulator.py", config, N+100)
+modulator.launch()
+
 # %% Launch Loop Class
 loop = hardwareLauncher("../pyRTC/Loop.py", config, N+4)
 loop.launch()
@@ -47,6 +47,9 @@ pidOptim = PIDOptimizer(read_yaml_file(config)["optimizer"]["pid"], loop)
 # %%
 from pyRTC.hardware.NCPAOptimizer import NCPAOptimizer
 ncpaOptim = NCPAOptimizer(read_yaml_file(config)["optimizer"]["ncpa"], loop, slopes)
+
+
+loopOptim = loopOptimizer(read_yaml_file(config)["optimizer"]["loop"], loop)
 
 # %% Calibrate
 if RECALIBRATE == True:
@@ -77,6 +80,7 @@ if RECALIBRATE == True:
     loop.setProperty("numItersIM", 100)
     loop.setProperty("IMFile", folder + "IM_PYWFS.npy")
     wfc.run("flatten")
+    time.sleep(1)
     loop.run("computeIM")
     loop.run("saveIM")
     wfc.run("flatten")
@@ -90,28 +94,35 @@ if RECALIBRATE == True:
     loop.setProperty("numItersIM", 50000)
     loop.setProperty("IMFile", folder + "IM_PYWFS_OL_docrime.npy")
     wfc.run("flatten")
+    time.sleep(1)
     loop.run("computeIM")
[…]
