+++ pyRTC/hardware/SpinnakerScienceCam.py
@@ -102,27 +102,5 @@ class spinCam(ScienceCamera):
 
 if __name__ == "__main__":
 
-    # Create argument parser
-    parser = argparse.ArgumentParser(description="Read a config file from the command line.")
-
-    # Add command-line argument for the config file
-    parser.add_argument("-c", "--config", required=True, help="Path to the config file")
-    parser.add_argument("-p", "--port", required=True, help="Port for communication")
-
-    # Parse command-line arguments
-    args = parser.parse_args()
-
-    conf = read_yaml_file(args.config)
-
-    pid = os.getpid()
-    set_affinity((conf["psf"]["affinity"])%os.cpu_count()) 
-    decrease_nice(pid)
-
-    psf = spinCam(conf=conf["psf"])
-    psf.start()
-
-    l = Listener(psf, port = int(args.port))
-    while l.running:
-        l.listen()
-        time.sleep(1e-3)
+    launchComponent(spinCam, "psf", start = True)
         
