### 2024-09-30 3ebf5fb
Small changes to requirements to fix examples notebook
