+++ pyRTC/WavefrontSensor.py
@@ -9,6 +9,55 @@ import matplotlib.pyplot as plt
 from numba import jit
 from sys import platform
 
+@jit(nopython=True, nogil=True, cache=True, fastmath=True)
+def downsample_int32_image_jit(image, N):
+    """
+    Numba-optimized function to downsample a 2D int32 NumPy array by a factor N, returning int32 output.
+
+    Parameters:
+    - image: 2D NumPy array of int32 with shape (H, W)
+    - N: int, downsampling factor
+
+    Returns:
+    - downsampled_image: 2D NumPy array of int32 with shape (H//N, W//N)
+    """
+    H, W = image.shape
+
+    # Calculate padding sizes if H or W is not divisible by N
+    pad_H = (-H) % N
+    pad_W = (-W) % N
+
+    # Pad the image if necessary to make dimensions divisible by N
+    if pad_H > 0 or pad_W > 0:
+        # Create a new array with zeros
+        H_padded = H + pad_H
+        W_padded = W + pad_W
+        image_padded = np.zeros((H_padded, W_padded), dtype=np.int32)
+        image_padded[:H, :W] = image
+    else:
+        image_padded = image
+        H_padded, W_padded = H, W
+
+    # Initialize the output array
+    out_H = H_padded // N
+    out_W = W_padded // N
+    downsampled_image = np.zeros((out_H, out_W), dtype=np.int32)
+
+    # Loop over the output array indices with Numba's parallel loops
+    for i in range(out_H):
+        for j in range(out_W):
+            # Compute the sum over the N x N block
+            sum_block = 0
+            for di in range(N):
+                for dj in range(N):
+                    sum_block += image_padded[i*N + di, j*N + dj]
+            # Compute the mean
+            mean_value = sum_block / (N * N)
+            # Round and cast to int32
+            downsampled_image[i, j] = np.int32(round(mean_value))
+
+    return downsampled_image
+
 class WavefrontSensor(pyRTCComponent):
     """
     A pyRTCComponent which represents a Wavefront Sensor (camera). This is a general class which is 
@@ -112,16 +161,20 @@ class WavefrontSensor(pyRTCComponent):
         self.height = setFromConfig(conf, "height", 1)
         self.darkCount = setFromConfig(conf, "darkCount", 1000)
         self.darkFile = setFromConfig(conf, "darkFile", "")
+        self.downsampleFactor = setFromConfig(conf, "downsampleFactor", 0)
 
-        self.imageShape = (self.width, self.height)
+        self.imageRawShape = [self.width, self.height]
         self.imageRawDType = np.uint16
         self.imageDType = np.int32
-
-        self.imageRaw = ImageSHM("wfsRaw", self.imageShape, self.imageRawDType, gpuDevice = self.gpuDevice, consumer=False)
+        self.imageShape = [self.width, self.height]
+        if self.downsampleFactor > 0:
+            self.imageShape[0] = self.imageShape[0] // self.downsampleFactor
+            self.imageShape[1] = self.imageShape[1] // self.downsampleFactor
+        self.imageRaw = ImageSHM("wfsRaw", self.imageRawShape, self.imageRawDType, gpuDevice = self.gpuDevice, consumer=False)
         self.image = ImageSHM("wfs", self.imageShape, self.imageDType, gpuDevice = self.gpuDevice, consumer=False)
 
         self.data = np.zeros(self.imageShape, dtype=self.imageRawDType)
-        self.dark = np.zeros(self.imageShape, dtype=self.imageDType)
+        self.dark = np.zeros(self.imageRawShape, dtype=self.imageDType)
 
         self.loadDark()
 
@@ -199,7 +252,12 @@ class WavefrontSensor(pyRTCComponent):
         Writes the current image data to shared memory. Both raw, and dark subtracted.
[…]
