+++ pyRTC/hardware/basicPredictLoop.py
@@ -1,0 +1,562 @@
+import numpy as np
+import torch
+import torch.nn as nn
+import torch.optim as optim
+from collections import deque
+from pyRTC.Loop import *
+from tqdm import tqdm
+from torch.utils.data import random_split, DataLoader
+
+
+class Discriminator(nn.Module):
+    def __init__(self, image_size):
+        super(Discriminator, self).__init__()
+        N, M = image_size  # Image dimensions
+
+        # Define the convolutional layers with smaller kernel sizes and strides
+        self.model = nn.Sequential(
+            nn.Conv2d(1, 64, kernel_size=3, stride=2, padding=1),  # Output: (batch, 64, N/2, M/2)
+            nn.LeakyReLU(0.2, inplace=True),
+
+            nn.Conv2d(64, 128, kernel_size=3, stride=2, padding=1),  # Output: (batch, 128, N/4, M/4)
+            nn.BatchNorm2d(128),
+            nn.LeakyReLU(0.2, inplace=True),
+
+            nn.Conv2d(128, 256, kernel_size=3, stride=2, padding=1),  # Output: (batch, 256, N/8, M/8)
+            nn.BatchNorm2d(256),
+            nn.LeakyReLU(0.2, inplace=True),
+
+            # Apply Adaptive Average Pooling to reduce spatial dimensions to (1, 1)
+            nn.AdaptiveAvgPool2d((1, 1)),
+        )
+
+        # Fully connected layer to output the probability
+        self.fc = nn.Sequential(
+            nn.Flatten(),
+            nn.Linear(256, 1),
+            nn.Sigmoid()
+        )
+
+    def forward(self, x):
+        # x shape: (batch_size, 1, N, M)
+        out = self.model(x)
+        out = self.fc(out)
+        return out
+
+
+class ConvLSTMCell(nn.Module):
+    """
+    Basic ConvLSTM cell.
+    """
+    def __init__(self, input_channels, hidden_channels, kernel_size, bias=True):
+        super(ConvLSTMCell, self).__init__()
+        
+        self.input_channels = input_channels
+        self.hidden_channels = hidden_channels
+        
+        self.kernel_size = kernel_size if isinstance(kernel_size, tuple) else (kernel_size, kernel_size)
+        self.padding = tuple(k // 2 for k in self.kernel_size)
+        self.bias = bias
+        
+        self.conv = nn.Conv2d(
+            in_channels=self.input_channels + self.hidden_channels,
+            out_channels=4 * self.hidden_channels,
+            kernel_size=self.kernel_size,
+            padding=self.padding,
+            bias=self.bias
+        )
+    
+    def forward(self, input_tensor, cur_state):
+        h_cur, c_cur = cur_state
+        
+        # Concatenate input and previous hidden state
+        combined = torch.cat([input_tensor, h_cur], dim=1)  # (batch, input_channels + hidden_channels, height, width)
+        
+        # Compute gates
+        combined_conv = self.conv(combined)
+        cc_i, cc_f, cc_o, cc_g = torch.split(combined_conv, self.hidden_channels, dim=1)
+        
+        # Input gate
+        i = torch.sigmoid(cc_i)
+        # Forget gate
+        f = torch.sigmoid(cc_f)
+        # Output gate
+        o = torch.sigmoid(cc_o)
+        # Cell gate
+        g = torch.tanh(cc_g)
+        
+        # New cell state
+        c_next = f * c_cur + i * g
+        # New hidden state
+        h_next = o * torch.tanh(c_next)
+        
+        return h_next, c_next
+    
+    def init_hidden(self, batch_size, height, width, device=None):
+        device = device or self.conv.weight.device
[…]
