+++ pyRTC/hardware/basicPredictLoop.py
@@ -242,6 +242,19 @@ class ConvLSTMModel(nn.Module):
 
 class basicPredictLoop(Loop):
     def __init__(self, conf) -> None:
+        self.T = conf["T"]
+        self.K = conf["K"]
+        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+        self.validSubApsFile = conf["validSubApsFile"]
+        self.validSubAps = np.load(self.validSubApsFile)
+        self.numXSlopes = int(np.sum(self.validSubAps) / 2)
+        self.numXSlopes2D = int(self.validSubAps.shape[1]/2)
+        self.slopemask = self.validSubAps[:,:self.numXSlopes2D]
+        self.history = np.zeros((self.K, *self.validSubAps.shape), dtype=np.float32)
+        self.curSignal2D = np.zeros_like(self.history[0])
+        self.history_GPU = torch.tensor(self.history, device = self.device)
+        self.s_pol = torch.zeros(np.sum(self.validSubAps), dtype=torch.float32, device=self.device)
+        self.s_pol_pred = torch.zeros(np.sum(self.validSubAps), dtype=torch.float32, device=self.device)
         #Initialize the pyRTC super class
         super().__init__(conf)
         """
@@ -257,8 +270,7 @@ class basicPredictLoop(Loop):
         - num_epochs (int): Number of epochs for training.
         - batch_size (int): Batch size for training.
         """
-        self.T = conf["T"]
-        self.K = conf["K"]
+
         self.hidden_size = conf["hidden_size"]
         self.num_layers = conf["num_layers"]
         self.learning_rate = conf["learning_rate"]
@@ -266,8 +278,8 @@ class basicPredictLoop(Loop):
         self.batch_size = conf["batch_size"]
         self.recordLength = 0
         # Move the model and its parameters to the GPU
-        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-        self.validSubAps = np.load("./calib/validSubAps.npy")
+        
+        
         # Define the LSTM-based predictive controller model
         model = ConvLSTMModel( 
                          hidden_channels=[64, 64], 
@@ -287,17 +299,28 @@ class basicPredictLoop(Loop):
         self.optimizer_D = torch.optim.Adam(self.discriminator.parameters(), lr=self.learning_rate, betas=(0.5, 0.999))
 
         # Buffer to hold the most recent K control vectors for real-time prediction
-        
-        self.slopemask = self.validSubAps[:,:self.validSubAps.shape[1]//2]
-        self.history = np.zeros((self.K, *self.validSubAps.shape), dtype=self.CM.dtype)
-        self.curSignal2D = np.zeros_like(self.history[0])
         self.polShm = ImageSHM("pol", self.curSignal2D.shape, self.curSignal2D.dtype, gpuDevice=self.gpuDevice, consumer=False)
         self.slopesBuffer = None
         self.predict = False
         self.numRecords = 0
         self.recordLength = 0
         self.record = False
-        self.gamma = 0.01
+        self.gamma = 0
+
+
+
+    def toDevice(self):
+        self.fIM_GPU = torch.tensor(self.fIM, device= self.device)
+        self.curSignal2D_GPU = torch.tensor(self.curSignal2D, device= self.device)
+        self.slopemask_GPU = torch.tensor(self.slopemask, device= self.device)
+        self.validSubAps_GPU = torch.tensor(self.validSubAps, device= self.device)
+        self.gCM_GPU = torch.tensor(self.gCM, device= self.device)
+        return
+    
+    def computeCM(self):
+        super().computeCM()
+        self.toDevice()
[…]
