+++ SHARP_LAB/config.yaml
@@ -44,7 +44,7 @@ wfs:
 slopes:
   type: SHWFS
   signalType: slopes
-  refSlopesFile: "/home/whetstone/pyRTC/SHARP_LAB/calib/ref.npy"
+  refSlopesFile: "/home/whetstone/pyRTC/SHARP_LAB/calib/ref_SH.npy"
   validSubApsFile: "/home/whetstone/pyRTC/SHARP_LAB/calib/validSubAps.npy"
   subApSpacing: 15.54 #7.78
   subApOffsetX: 13
@@ -86,8 +86,8 @@ wfc:
   serial: "BAX118"
   numActuators: 97
   numModes: 94
-  flatFile: "/home/whetstone/pyRTC/SHARP_LAB/calib/wfcShape.npy"
-  saveFile: "/home/whetstone/pyRTC/SHARP_LAB/calib/wfcShape.npy"
+  flatFile: "/home/whetstone/pyRTC/SHARP_LAB/calib/wfcShape_SH.npy"
+  saveFile: "/home/whetstone/pyRTC/SHARP_LAB/calib/wfcShape_SH.npy"
   m2cFile:  "/home/whetstone/pyRTC/SHARP_LAB/calib/m2c_kl.npy" 
   floatingActuatorsFile: ""
   floatingInfluenceRadius: 1
@@ -109,7 +109,7 @@ psf:
   height: 128
   darkCount: 2000
   darkFile: "/home/whetstone/pyRTC/SHARP_LAB/calib/psfDark.npy"
-  modelFile: "/home/whetstone/pyRTC/SHARP_LAB/calib/modelPSF.npy"
+  modelFile: "/home/whetstone/pyRTC/SHARP_LAB/calib/modelPSF_SH.npy"
   integration: 1
   affinity: 6
   functions:

[…]
