+++ SHARP_LAB/basicPredictRTC.py
@@ -1,0 +1,199 @@
+# %% Imports
+import matplotlib.pyplot as plt
+import time
+from tqdm import tqdm
+#%%
+import sys
+tmp = sys.stdout
+from pyRTC import *
+RECALIBRATE = False
+sys.stdout = tmp
+import logging
+import matplotlib
+logging.getLogger('matplotlib').setLevel(logging.WARNING)
+#%%
+config = '/home/whetstone/pyRTC/SHARP_LAB/config_basic_predict_pywfs.yaml'
+N = np.random.randint(3000,6000)
+folder = "/home/whetstone/pyRTC/SHARP_LAB/calib/"
+conf = utils.read_yaml_file(config)
+
+# %% Launch WFS
+wfs = hardwareLauncher("../pyRTC/hardware/ximeaWFS.py", config, N+1)
+wfs.launch()
+
+#%% Create Slope computation 
+slopes = hardwareLauncher("../pyRTC/SlopesProcess.py", config, N+2)
+slopes.launch()
+
+# %% Launch DM
+wfc = hardwareLauncher("../pyRTC/hardware/ALPAODM.py", config, N)
+wfc.launch()
+
+# %% Initialize our AO loop object
+loop = hardwareLauncher("../pyRTC/hardware/basicPredictLoop.py", config, N+4)
+loop.launch()
+# conf = utils.read_yaml_file(config)
+# confLoop = conf["loop"]
+# from pyRTC.hardware.basicPredictLoop import basicPredictLoop
+# loop = basicPredictLoop(conf=confLoop)
+
+# %% Launch PSF Cam
+psfCam = hardwareLauncher("../pyRTC/hardware/SpinnakerScienceCam.py", config, N+10)
+psfCam.launch()
+
+# %% Launch Modulator
+modulator = hardwareLauncher("../pyRTC/hardware/PIModulator.py", config, N+100)
+modulator.launch()
+
+#%%Adjust frame delay
+psfCam.setProperty("integrationLength", 50)
+# wfc.run("setDelay",loop.T)#loop.getProperty("T"))
+wfc.run("flatten")  
+# loop.predict = False
+# loop.numDroppedModes = 9
+# loop.gain = 0.3
+# loop.leakyGain =  0.01
+# loop.loadIM()
+# time.sleep(1)
+# loop.start()
+
+wfc.run("setDelay",loop.getProperty("T"))
+loop.setProperty("predict", False)
+loop.setProperty("numDroppedModes", 9)
+loop.setProperty("gain",0.3)
+loop.setProperty("leakyGain", 0.01)
+loop.run("loadIM")
+time.sleep(1)
+loop.run("start")
+# %%
+loop.listen(int(2**14))
+loop.stop()
+# loop.run("listen", int(2**10))
+# # loop.slopesBuffer = np.load("./calib/slopesBuffer.npy")
+# loop.run("stop") 
+# %%
+loop.num_epochs = 100
+loop.train()
+
+# loop.run("loadModels")
+# loop.setProperty("predict", True)
+# %%
+loop.run("start") 
+loop.run("setGain", 0.4)
+loop.predict=False
+time.sleep(5)
+loop.gamma = 0.6
+# loop.predict=True
+#%%
+def recordStrehl(strehlShm, N=10):
+    val = 0
+    for j in range(N):
+        val += np.mean(strehlShm.read())
+    return val/N
+
+def resetLoop(hardReal = False):
+    if hardReal:
+        loop.run("stop")
+        loop.run("flatten")
+        time.sleep(0.5)
+        loop.run("start") 
+        time.sleep(0.5)
+    else:
+
+        loop.stop() 
+        loop.flatten()
[…]

[…]
