### 2024-10-09 28c11ed
Changed sim to have faster relative wind, got results from predictive control on sim
